Copyedit and update for the ROM Transfer Tutorial
As this is only my second pull request I still have to finish learning the markdown system used here, but following the example set my the existing text I think I formatted the links and such as they should be. The main changes I made were a rewrite of the text before the tutorial to improve clarity and structure, as well as a lesser rewrite of the portion of the tutorial recommending the use of the FunKey Wiki and LibRetro box art databases (with the link for the FunKey Wiki box art database updated to the new page for it). After I'm done updating pages of the wiki in the next few days, I'll see which other pages I can perhaps help to improve.